Suppress empty IN-predicate warning when running tests.
This is the full text of the warning, run because get_most_complete_profiles
is being called on an empty db:

sqlalchemy/sql/default_comparator.py:153: SAWarning: The IN-predicate
on "users.id" was invoked with an empty sequence. This results in
a contradiction, which nonetheless can be expensive to evaluate.
Consider alternative strategies for improved performance.